Elasticsearch: add version 7 compatibility
* new bulk api version
* test with testcontainers
