+++ elasticsearch/src/main/scala/akka/stream/alpakka/elasticsearch/impl/RestBulkApiV7.scala
@@ -14,45 +14,35 @@ import scala.collection.immutable
 /**
  * Internal API.
  *
- * REST API implementation for some Elasticsearch 5 version.
- * https://www.elastic.co/guide/en/elasticsearch/reference/5.6/docs-bulk.html
+ * REST API implementation for some Elasticsearch 7 version.
+ * https://www.elastic.co/guide/en/elasticsearch/reference/7.6/docs-bulk.html
  */
 @InternalApi
-private[impl] final class RestBulkApiV5[T, C](indexName: String,
-                                              typeName: String,
+private[impl] final class RestBulkApiV7[T, C](indexName: String,
                                               versionType: Option[String],
                                               messageWriter: MessageWriter[T])
     extends RestBulkApi[T, C] {
-
-  private lazy val typeNameTuple = "_type" -> JsString(typeName)
 
   def toJson(messages: immutable.Seq[WriteMessage[T, C]]): String =
     messages
       .map { message =>
         val sharedFields: Seq[(String, JsString)] = Seq(
-            "_index" -> JsString(message.indexName.getOrElse(indexName)),
-            typeNameTuple
+            "_index" -> JsString(message.indexName.getOrElse(indexName))
           ) ++ message.customMetadata.map { case (field, value) => field -> JsString(value) }
         val tuple: (String, JsObject) = message.operation match {
           case Index =>
             val fields = Seq(
-              optionalNumber(message.version, "_version"),
+              optionalNumber(message.version, "version"),
               optionalString(versionType, "version_type"),
               optionalString(message.id, "_id")
             ).flatten
             "index" -> JsObject(sharedFields ++ fields: _*)
           case Create => "create" -> JsObject(sharedFields ++ optionalString(message.id, "_id"): _*)
-          case Update | Upsert =>
-            val fields =
-              ("_id" -> JsString(message.id.get)) +: Seq(
-                optionalNumber(message.version, "_version"),
-                optionalString(versionType, "version_type")
-              ).flatten
-            "update" -> JsObject(sharedFields ++ fields: _*)
+          case Update | Upsert => "update" -> JsObject(sharedFields :+ ("_id" -> JsString(message.id.get)): _*)
           case Delete =>
             val fields =
               ("_id" -> JsString(message.id.get)) +: Seq(
-                optionalNumber(message.version, "_version"),
+                optionalNumber(message.version, "version"),
                 optionalString(versionType, "version_type")
               ).flatten
             "delete" -> JsObject(sharedFields ++ fields: _*)
