+++ elasticsearch/src/test/java/docs/javadsl/ElasticsearchTest.java
@@ -7,12 +7,10 @@ package docs.javadsl;
 import akka.Done;
 import akka.NotUsed;
 import akka.actor.ActorSystem;
-import akka.japi.Pair;
 import akka.stream.ActorMaterializer;
 // #init-client
 import akka.stream.alpakka.elasticsearch.*;
 import akka.stream.alpakka.elasticsearch.javadsl.*;
-
 // #init-client
 import akka.stream.alpakka.testkit.javadsl.LogCapturingJunit4;
 import akka.stream.javadsl.Sink;
@@ -22,12 +20,14 @@ import akka.testkit.javadsl.TestKit;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.http.entity.StringEntity;
 import org.apache.http.message.BasicHeader;
-import org.codelibs.elasticsearch.runner.ElasticsearchClusterRunner;
 // #init-client
-import org.elasticsearch.client.RestClient;
 import org.apache.http.HttpHost;
+import org.elasticsearch.client.RestClient;
 // #init-client
 import org.junit.*;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+import org.testcontainers.elasticsearch.ElasticsearchContainer;
 
 import java.io.IOException;
 import java.time.Duration;
@@ -38,13 +38,26 @@ import java.util.stream.Collectors;
 
 import static org.junit.Assert.assertEquals;
 
+@RunWith(value = Parameterized.class)
 public class ElasticsearchTest {
   @Rule public final LogCapturingJunit4 logCapturing = new LogCapturingJunit4();
 
-  private static ElasticsearchClusterRunner runner;
+  @Parameterized.Parameters(name = "{index}: version={0} api={1}")
+  public static Iterable<Object[]> data() {
+    return Arrays.asList(
+        new Object[][] {
+          {"6.8.0", ApiVersion.V5},
+          {"7.6.0", ApiVersion.V7}
+        });
+  }
+
+  private static ApiVersion apiVersion;
+  private static ElasticsearchContainer container;
   private static RestClient client;
   private static ActorSystem system;
   private static ActorMaterializer materializer;
+
+  public ElasticsearchTest(String esVersion, ApiVersion apiVersion) {}
 
   // #define-class
   public static class Book {
@@ -58,27 +71,16 @@ public class ElasticsearchTest {
   }
   // #define-class
 
-  @BeforeClass
-  public static void setup() throws IOException {
-    runner = new ElasticsearchClusterRunner();
-    runner.build(
-        ElasticsearchClusterRunner.newConfigs()
-            .baseHttpPort(9200)
-            .baseTransportPort(9300)
-            .numOfNode(1)
-            .disableESLogger());
-    runner.ensureYellow();
-
+  @Parameterized.BeforeParam
+  public static void beforeParam(String esVersion, ApiVersion esApiVersion) throws IOException {
+    apiVersion = esApiVersion;
+    container =
+        new ElasticsearchContainer("docker.elastic.co/elasticsearch/elasticsearch:" + esVersion);
+    container.start();
     // #init-client
 
-    client = RestClient.builder(new HttpHost("localhost", 9201)).build();
+    client = RestClient.builder(HttpHost.create(container.getHttpHostAddress())).build();
     // #init-client
-
-    // #init-mat
-    system = ActorSystem.create();
-    materializer = ActorMaterializer.create(system);
-    // #init-mat
-
     register("source", "Akka in Action");
     register("source", "Programming in Scala");
     register("source", "Learning Scala");
@@ -86,13 +88,20 @@ public class ElasticsearchTest {
     register("source", "Scala Puzzlers");
     register("source", "Effective Akka");
     register("source", "Akka Concurrency");
-    flush("source");
+    flushAndRefresh("source");
+  }
+
+  @BeforeClass
+  public static void setup() throws IOException {
+    // #init-mat
+    system = ActorSystem.create();
+    materializer = ActorMaterializer.create(system);
+    // #init-mat
   }
 
   @AfterClass
   public static void teardown() throws Exception {
-    runner.close();
-    runner.clean();
+    container.close();
     client.close();
     TestKit.shutdownActorSystem(system);
   }
@@ -102,8 +111,9 @@ public class ElasticsearchTest {
     StreamTestKit.assertAllStagesStopped(materializer);
   }
 
-  private static void flush(String indexName) throws IOException {
+  private static void flushAndRefresh(String indexName) throws IOException {
     client.performRequest("POST", indexName + "/_flush");
+    client.performRequest("POST", indexName + "/_refresh");
   }
 
   private static void register(String indexName, String title) throws IOException {
@@ -116,6 +126,9 @@ public class ElasticsearchTest {
   }
 
   private void documentation() {
+    // #init-client
+    client = RestClient.builder(new HttpHost("localhost", 9201)).build();
+    // #init-client
     // #source-settings
     ElasticsearchSourceSettings sourceSettings =
         ElasticsearchSourceSettings.create().withBufferSize(10);
@@ -125,7 +138,8 @@ public class ElasticsearchTest {
         ElasticsearchWriteSettings.create()
             .withBufferSize(10)
             .withVersionType("internal")
-            .withRetryLogic(RetryAtFixedRate.create(5, Duration.ofSeconds(1)));
+            .withRetryLogic(RetryAtFixedRate.create(5, Duration.ofSeconds(1)))
+            .withApiVersion(ApiVersion.V5);
     // #sink-settings
   }
 
@@ -134,7 +148,8 @@ public class ElasticsearchTest {
     // Copy source/book to sink1/book through JsObject stream
     // #run-jsobject
     ElasticsearchSourceSettings sourceSettings = ElasticsearchSourceSettings.create();
-    ElasticsearchWriteSettings sinkSettings = ElasticsearchWriteSettings.create();
+    ElasticsearchWriteSettings sinkSettings =
+        ElasticsearchWriteSettings.create().withApiVersion(apiVersion);
 
     Source<ReadResult<Map<String, Object>>, NotUsed> source =
         ElasticsearchSource.create("source", "_doc", "{\"match_all\": {}}", sourceSettings, client);
@@ -148,7 +163,7 @@ public class ElasticsearchTest {
 
     f1.toCompletableFuture().get();
 
-    flush("sink1");
+    flushAndRefresh("sink1");
 
     // Assert docs in sink1/book
     CompletionStage<List<String>> f2 =
@@ -182,7 +197,8 @@ public class ElasticsearchTest {
     // Copy source/book to sink2/book through JsObject stream
     // #run-typed
     ElasticsearchSourceSettings sourceSettings = ElasticsearchSourceSettings.create();
-    ElasticsearchWriteSettings sinkSettings = ElasticsearchWriteSettings.create();
+    ElasticsearchWriteSettings sinkSettings =
+        ElasticsearchWriteSettings.create().withApiVersion(apiVersion);
 
     Source<ReadResult<Book>, NotUsed> source =
         ElasticsearchSource.typed(
@@ -197,7 +213,7 @@ public class ElasticsearchTest {
 
     f1.toCompletableFuture().get();
 
-    flush("sink2");
+    flushAndRefresh("sink2");
 
     // Assert docs in sink2/book
     CompletionStage<List<String>> f2 =
@@ -244,14 +260,16 @@ public class ElasticsearchTest {
                 ElasticsearchFlow.create(
                     "sink3",
                     "_doc",
-                    ElasticsearchWriteSettings.create().withBufferSize(5),
+                    ElasticsearchWriteSettings.create()
+                        .withApiVersion(apiVersion)
+                        .withBufferSize(5),
                     client,
                     new ObjectMapper()))
             .runWith(Sink.seq(), materializer);
     // #run-flow
 
     List<WriteResult<Book, NotUsed>> result1 = f1.toCompletableFuture().get();
-    flush("sink3");
+    flushAndRefresh("sink3");
 
     for (WriteResult<Book, NotUsed> aResult1 : result1) {
       assertEquals(true, aResult1.success());
@@ -301,14 +319,16 @@ public class ElasticsearchTest {
                 ElasticsearchFlow.create(
                     indexName,
                     "_doc",
-                    ElasticsearchWriteSettings.create().withBufferSize(5),
+                    ElasticsearchWriteSettings.create()
+                        .withApiVersion(apiVersion)
+                        .withBufferSize(5),
                     client,
                     StringMessageWriter.getInstance()))
             .runWith(Sink.seq(), materializer);
     // #string
 
     List<WriteResult<String, NotUsed>> result1 = write.toCompletableFuture().get();
-    flush(indexName);
+    flushAndRefresh(indexName);
 
     for (WriteResult<String, NotUsed> aResult1 : result1) {
       assertEquals(true, aResult1.success());
@@ -348,13 +368,17 @@ public class ElasticsearchTest {
     Source.from(requests)
         .via(
             ElasticsearchFlow.create(
-                "sink8", "_doc", ElasticsearchWriteSettings.create(), client, new ObjectMapper()))
+                "sink8",
+                "_doc",
+                ElasticsearchWriteSettings.create().withApiVersion(apiVersion),
+                client,
+                new ObjectMapper()))
         .runWith(Sink.seq(), materializer)
         .toCompletableFuture()
         .get();
     // #multiple-operations
 
-    flush("sink8");
+    flushAndRefresh("sink8");
 
     // Assert docs in sink8/book
     CompletionStage<List<String>> f2 =
@@ -405,7 +429,9 @@ public class ElasticsearchTest {
                 ElasticsearchFlow.createWithPassThrough(
                     "sink6",
                     "_doc",
-                    ElasticsearchWriteSettings.create().withBufferSize(5),
+                    ElasticsearchWriteSettings.create()
+                        .withApiVersion(apiVersion)
+                        .withBufferSize(5),
                     client,
                     new ObjectMapper()))
             .map(
@@ -419,7 +445,7 @@ public class ElasticsearchTest {
             .runWith(Sink.ignore(), materializer);
     // #kafka-example
     kafkaToEs.toCompletableFuture().get(5, TimeUnit.SECONDS); // Wait for it to complete
-    flush("sink6");
+    flushAndRefresh("sink6");
 
     // Make sure all messages was committed to kafka
     assertEquals(Arrays.asList(0, 1, 2), kafkaCommitter.committedOffsets);
@@ -458,14 +484,14 @@ public class ElasticsearchTest {
             ElasticsearchFlow.create(
                 indexName,
                 typeName,
-                ElasticsearchWriteSettings.create().withBufferSize(5),
+                ElasticsearchWriteSettings.create().withApiVersion(apiVersion).withBufferSize(5),
                 client,
                 new ObjectMapper()))
         .runWith(Sink.seq(), materializer)
         .toCompletableFuture()
         .get();
 
-    flush(indexName);
+    flushAndRefresh(indexName);
 
     // Search document and assert it having version 1
     ReadResult<Book> message =
@@ -482,7 +508,7 @@ public class ElasticsearchTest {
 
     assertEquals(1L, message.version().get());
 
-    flush(indexName);
+    flushAndRefresh(indexName);
 
     // Update document to version 2
     Source.single(WriteMessage.createIndexMessage("1", book).withVersion(1L))
@@ -490,14 +516,17 @@ public class ElasticsearchTest {
             ElasticsearchFlow.create(
                 indexName,
                 typeName,
-                ElasticsearchWriteSettings.create().withBufferSize(5),
+                ElasticsearchWriteSettings.create()
+                    .withApiVersion(apiVersion)
+                    .withBufferSize(5)
+                    .withVersionType("external"),
                 client,
                 new ObjectMapper()))
         .runWith(Sink.seq(), materializer)
         .toCompletableFuture()
         .get();
 
-    flush(indexName);
+    flushAndRefresh(indexName);
 
     // Try to update document with wrong version to assert that we can send it
     long oldVersion = 1;
@@ -507,7 +536,10 @@ public class ElasticsearchTest {
                 ElasticsearchFlow.create(
                     indexName,
                     typeName,
-                    ElasticsearchWriteSettings.create().withBufferSize(5),
+                    ElasticsearchWriteSettings.create()
+                        .withApiVersion(apiVersion)
+                        .withBufferSize(5)
+                        .withVersionType("external"),
                     client,
                     new ObjectMapper()))
             .runWith(Sink.seq(), materializer)
@@ -534,14 +566,17 @@ public class ElasticsearchTest {
             ElasticsearchFlow.create(
                 indexName,
                 typeName,
-                ElasticsearchWriteSettings.create().withBufferSize(5).withVersionType("external"),
+                ElasticsearchWriteSettings.create()
+                    .withApiVersion(apiVersion)
+                    .withBufferSize(5)
+                    .withVersionType("external"),
                 client,
                 new ObjectMapper()))
         .runWith(Sink.seq(), materializer)
         .toCompletableFuture()
         .get();
 
-    flush(indexName);
+    flushAndRefresh(indexName);
 
     // Assert that the document's external version is saved
     ReadResult<Book> message =
@@ -599,14 +634,14 @@ public class ElasticsearchTest {
             ElasticsearchFlow.create(
                 indexName,
                 typeName,
-                ElasticsearchWriteSettings.create().withBufferSize(5),
+                ElasticsearchWriteSettings.create().withApiVersion(apiVersion).withBufferSize(5),
                 client,
                 new ObjectMapper()))
         .runWith(Sink.seq(), materializer)
         .toCompletableFuture()
         .get();
 
-    flush(indexName);
+    flushAndRefresh(indexName);
 
     // #custom-search-params
     // Search for docs and ask elastic to only return some fields
@@ -632,7 +667,7 @@ public class ElasticsearchTest {
             .toCompletableFuture()
             .get();
     // #custom-search-params
-    flush(indexName);
+    flushAndRefresh(indexName);
 
     assertEquals(
         docs.size(),
